requirements.txt
notebook
ipywidgets
tqdm
python-dotenv
aiogram
openai
